requirements.txt
python-dotenv
PyPDF2
langchain
openai
faiss-cpu
streamlit
jupyterlab
pycryptodome
tiktoken
